requirements.txt
mixpanel==4.10.1
email-validator==1.1.3
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
emoji==2.10.1
exceptiongroup==1.2.0
fastapi==0.95.2
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.6
pydantic==1.10.13
PyJWT==1.7.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
alembic==1.12.0
FastAPI-SQLAlchemy==0.2.1
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2==2.9.9
pydantic_core==2.14.6
SQLAlchemy==2.0.23
boto3==1.34.3
botocore==1.34.3
jmespath==1.0.1
python-dateutil==2.8.2
redis==4.5.5
s3transfer==0.9.0
six==1.16.0
urllib3==1.26.18
python-multipart==0.0.6
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gunicorn==21.2.0
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
openai==1.6.0
json5==0.9.14
PyYAML~=6.0.1
stripe~=7.10.0
celery~=5.3.6
newrelic==9.3.0
pusher==3.3.2
Jinja2==3.1.3
sendgrid==6.11.0
google-api-python-client==2.113.0
pytz==2024.1
rich==13.7.0
asgi-correlation-id[celery]==4.3.0
zerobouncesdk==1.1.0
confluent-kafka==2.3.0
paho-mqtt==1.6.1
Pillow==10.2.0
flower==2.0.1
python-jose==3.3.0
cryptography==43.0.0
bcrypt==4.2.0